feat: separate k8s tests
